Reviews have now been edited to adhere for what is shown on an individual profile and what is shown to other users via a class
